requirements.txt
python-binance
pandas
pandas_datareader
matplotlib
python-dotenv
yfinance